Use floats instead of doubles. Create FloatMatrix class, a 2D matrix …

+++ src/main/java/mp_pprl/core/graph/Edge.java
@@ -2,13 +2,13 @@ package mp_pprl.core.graph;
 
 import java.util.Objects;
 
-public record Edge(Cluster c1, Cluster c2, double metric) {
+public record Edge(Cluster c1, Cluster c2, float metric) {
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Edge edge = (Edge) o;
-        return Double.compare(metric, edge.metric) == 0 && Objects.equals(c1, edge.c1) && Objects.equals(c2, edge.c2);
+        return Float.compare(metric, edge.metric) == 0 && Objects.equals(c1, edge.c1) && Objects.equals(c2, edge.c2);
     }
 
     @Override
